try another approach: making a static class member of type boolean to hold the result of monitoring
