+++ RVM/analysis/LogMonitor.java
@@ -35,7 +35,7 @@ public class LogMonitor {
         for (String eventName : TableCol.keySet()) {
             int[] types4CurTable = TableCol.get(eventName);
 
-            Class[] argTyList4CurMeth = new Class[types4CurTable.length + 1];
+            Class[] argTyList4CurMeth = new Class[types4CurTable.length];
             for (int i = 0; i < types4CurTable.length; i++) {
                 switch (types4CurTable[i]) {
                     case RegHelper.INT_TYPE:
@@ -57,9 +57,6 @@ public class LogMonitor {
                     }
                 }
             }
-
-            //append the long type at the end which is the type for the timestamp.
-            argTyList4CurMeth[types4CurTable.length] = boolean[].class; //the array for holding violation result
 
             if (FormulaExtractor.monitoredEventList.contains(eventName)) {
                 String methName = eventName + "Event";
