+++ test/insert-smallLog/insert.rvm
@@ -1,12 +1,11 @@
 package rvm;
-  Insert(String user,String db) {
- // String user;
- // String db;
+Insert(String user,String db) {
+public static boolean hasViolation;
 
-  event insert (String user,String db,String p,String data, boolean[] result) {
+  event insert (String user,String db,String p,String data) {
 		if(db.equals("db2") && !user.equals("script1"))
 		{
-			result[0] = true;
+			hasViolation = true;
 		}
 	}
 }
